Update Source interface to use SourceID and JobID types
The previous implementation used int64 for both, which can be mixed up
easily. Using distinct types adds a layer of type safety checked by the
compiler.

+++ pkg/engine/filesystem.go
@@ -29,7 +29,7 @@ func (e *Engine) ScanFileSystem(ctx context.Context, c sources.FilesystemConfig)
 
 	fileSystemSource := &filesystem.Source{}
 	fileSystemSource.WithFilter(c.Filter)
-	if err := fileSystemSource.Init(ctx, sourceName, int64(jobID), int64(sourceID), true, &conn, runtime.NumCPU()); err != nil {
+	if err := fileSystemSource.Init(ctx, sourceName, jobID, sourceID, true, &conn, runtime.NumCPU()); err != nil {
 		return err
 	}
 	_, err = e.sourceManager.Run(ctx, sourceName, fileSystemSource)
