+++ pkg/engine/s3.go
@@ -62,7 +62,7 @@ func (e *Engine) ScanS3(ctx context.Context, c sources.S3Config) error {
 	sourceID, jobID, _ := e.sourceManager.GetIDs(ctx, sourceName, s3.SourceType)
 
 	s3Source := &s3.Source{}
-	if err := s3Source.Init(ctx, sourceName, int64(jobID), int64(sourceID), true, &conn, runtime.NumCPU()); err != nil {
+	if err := s3Source.Init(ctx, sourceName, jobID, sourceID, true, &conn, runtime.NumCPU()); err != nil {
 		return err
 	}
 	_, err = e.sourceManager.Run(ctx, sourceName, s3Source)
